+++ bot/src/main/kotlin/io/github/populus_omnibus/vikbot/api/CustomMessageData.kt
@@ -1,0 +1,4 @@
+package io.github.populus_omnibus.vikbot.api
+import net.dv8tion.jda.api.entities.Message
+
+abstract class CustomMessageData(val msg: Message)
