+++ bot/src/main/kotlin/io/github/populus_omnibus/vikbot/bot/modules/RoleSelector.kt
@@ -3,17 +3,13 @@ package io.github.populus_omnibus.vikbot.bot.modules
 import io.github.populus_omnibus.vikbot.VikBotHandler
 import io.github.populus_omnibus.vikbot.VikBotHandler.config
 import io.github.populus_omnibus.vikbot.VikBotHandler.jda
-import io.github.populus_omnibus.vikbot.api.EventResult
+import io.github.populus_omnibus.vikbot.api.*
 import io.github.populus_omnibus.vikbot.api.annotations.Module
 import io.github.populus_omnibus.vikbot.api.commands.CommandGroup
 import io.github.populus_omnibus.vikbot.api.commands.SlashCommand
 import io.github.populus_omnibus.vikbot.api.commands.SlashOptionType
 import io.github.populus_omnibus.vikbot.api.commands.adminOnly
-import io.github.populus_omnibus.vikbot.api.createMemory
 import io.github.populus_omnibus.vikbot.api.interactions.IdentifiableInteractionHandler
-import io.github.populus_omnibus.vikbot.api.maintainEvent
-import io.github.populus_omnibus.vikbot.api.plusAssign
-import io.github.populus_omnibus.vikbot.bot.CustomMessageData
 import io.github.populus_omnibus.vikbot.bot.RoleEntry
 import io.github.populus_omnibus.vikbot.bot.RoleEntry.RoleDescriptor
 import io.github.populus_omnibus.vikbot.bot.RoleGroup
@@ -37,6 +33,31 @@ import net.dv8tion.jda.api.interactions.components.selections.StringSelectMenu
 import net.dv8tion.jda.api.utils.messages.MessageCreateBuilder
 import net.dv8tion.jda.api.utils.messages.MessageEditBuilder
 import org.slf4j.kotlin.getLogger
+import kotlin.collections.List
+import kotlin.collections.Map
+import kotlin.collections.any
+import kotlin.collections.asSequence
+import kotlin.collections.elementAtOrNull
+import kotlin.collections.filter
+import kotlin.collections.filterIsInstance
+import kotlin.collections.find
+import kotlin.collections.firstOrNull
+import kotlin.collections.forEach
+import kotlin.collections.getOrNull
+import kotlin.collections.getOrPut
+import kotlin.collections.intersect
+import kotlin.collections.joinToString
+import kotlin.collections.map
+import kotlin.collections.mapNotNull
+import kotlin.collections.minus
+import kotlin.collections.mutableListOf
+import kotlin.collections.plusAssign
+import kotlin.collections.set
+import kotlin.collections.sortedBy
+import kotlin.collections.take
+import kotlin.collections.toMutableList
+import kotlin.collections.toSet
+import kotlin.collections.toSortedMap
 import kotlin.time.Duration.Companion.minutes
